Move facet spec to `spec/facet.ts`
